requirements.txt
configparser==5.0.1
docutils==0.16
lockfile==0.12.2
mysql-connector-python==8.0.23
protobuf==3.14.0
python-daemon==2.2.4
scapy==2.4.4
six==1.15.0
